requirements_local.txt
fastapi
uvicorn[standard]
gunicorn
httpx
loguru
sqlalchemy
psycopg2-binary
boto3
requests
python-multipart
pathlib
prometheus-client
PyJWT
cryptography
pyjwt[crypto]
python-logging-loki
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-exporter-otlp
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-exporter-otlp-proto-http
loguru
opentelemetry-instrumentation-logging
redis[hiredis]>=5.0.0
pymysql
python-dotenv
clerk_backend_api
passlib
bcrypt
cryptography
brotli
fastapi_mcp
sqlalchemy-libsql
authlib
itsdangerous
jinja2
openai-agents
twilio
sendgrid
beautifulsoup4
duckduckgo-search
